requirements.txt
defusedxml==0.6.0
requests>=2.23.0
python-dateutil>=2.8.1
boto3>=1.12.21
lxml>=4.5.0
